requirements.txt
cassandra-driver==3.13.0
pkg-resources==0.0.0
six==1.11.0
